Need to decide on sync semantics

+++ rplugin/python3/qualia/plugin.py
@@ -1,17 +1,21 @@
 from time import sleep
-from typing import Optional
+from typing import Optional, Type
 
 from orderedset import OrderedSet
 from pynvim import plugin, Nvim, command, attach, function
 
 from qualia.config import _FZF_LINE_DELIMITER, NVIM_DEBUG_PIPE
-from qualia.database import Database
+from qualia.database import Database, MuDatabase
 from qualia.driver import PluginDriver
-from qualia.models import NodeId
+from qualia.models import NodeId, MinimalDb, MutableDb, View, SourceId
+from qualia.services.file_db import FileDb
+from qualia.services.listener import RpcListenExternal
 from qualia.services.search import matching_nodes_content, fzf_input_line
 from qualia.utils.bootstrap_utils import bootstrap
-from qualia.utils.common_utils import exception_traceback, normalized_search_prefixes, live_logger
+from qualia.utils.common_utils import exception_traceback, normalized_search_prefixes, live_logger, StartLoggedThread
 from qualia.utils.plugin_utils import get_orphan_node_ids, PluginUtils
+
+Sources: dict[str, Type[MinimalDb]] = {'directory': FileDb}
 
 
 @plugin
@@ -19,6 +23,7 @@ class Qualia(PluginDriver):
     def __init__(self, nvim: Nvim, ide_debugging: bool = False):
         try:
             live_logger.attach_nvim(nvim)
+            StartLoggedThread(target=RpcListenExternal, name="RPClistener", delay_seconds=2)
             bootstrap()
             super().__init__(nvim, ide_debugging)
         except Exception as e:
@@ -29,19 +34,29 @@ class Qualia(PluginDriver):
     def trigger_sync_cmd(self, args: list = None) -> None:
         self.trigger_sync(True if args and int(args[0]) else False)
 
-    @command("NavigateNode", sync=True, nargs='?')
+    @command("NavigateNode", sync=True, nargs='*')
     def navigate_cur_node(self, args: list[NodeId] = None) -> None:
-        node_id = (args and args[0]) or self.line_info(self.current_line_number()).node_id
-        with Database() as db:
-            self.navigate_node(node_id, False, db)
+        view = self.line_node_view(self.current_line_number())
+        view.sub_tree = None
+
+        if args:
+            view.node_id = args[0]
+            if len(args) == 2:
+                view.source_id = args[1]
+            else:
+                live_logger.critical("The number of arguments should be less than equal to two. First optional argument is node ID and optional second is source ID.")
+                return
+
+        self.navigate_node(view, False, db)
 
     @command("HoistNode", sync=True)
     def hoist_node(self) -> None:
         line_num = self.current_line_number()
         view = self.line_node_view(line_num)
         with Database() as db:
-            db.set_node_view(view, self.file_path_transposed(self.nvim.current.buffer.name))
-            self.navigate_node(view.main_id, True, db)
+            if isinstance(db, MutableDb):
+                db.set_node_view(view)
+        self.navigate_node(view, True, db)
 
     @command("ToggleQualia", sync=True)
     def toggle_parser(self) -> None:
@@ -70,13 +85,16 @@ class Qualia(PluginDriver):
         grandparent_id = parent_line_info.parent_view.main_id
 
         with Database() as db:
-            cur_node_siblings = db.get_node_descendants(parent_id, transposed, False)
-            cur_node_siblings.remove(cur_node_id)
-            db.set_node_descendants(parent_id, cur_node_siblings, transposed)
+            if isinstance(db, MutableDb):
+                cur_node_siblings = db.get_node_descendants(parent_id, transposed, False, temporary)
+                cur_node_siblings.remove(cur_node_id)
+                db.set_node_descendants(parent_id, cur_node_siblings, transposed)
 
-            parent_node_siblings_list = list(db.get_node_descendants(grandparent_id, transposed, False))
-            parent_node_siblings_list.insert(parent_node_siblings_list.index(parent_id) + 1, cur_node_id)
-            db.set_node_descendants(grandparent_id, OrderedSet(parent_node_siblings_list), transposed)
+                parent_node_siblings_list = list(db.get_node_descendants(grandparent_id, transposed, False, temporary))
+                parent_node_siblings_list.insert(parent_node_siblings_list.index(parent_id) + 1, cur_node_id)
+                db.set_node_descendants(grandparent_id, OrderedSet(parent_node_siblings_list), transposed)
+            else:
+                PluginUtils.unmutable_db_error()
 
             # Preserve expanded state: restore view
 
@@ -115,25 +133,25 @@ class Qualia(PluginDriver):
 
     @command("TransposeNode", sync=True, nargs='?')
     def transpose(self, args: list[str] = None) -> None:
-        currently_transposed = self.file_path_transposed(self.nvim.current.buffer.name)
-        node_id = self.line_info(self.current_line_number()).node_id
         try:
             replace_buffer = False if args and int(args[0]) else True
         except ValueError:
             live_logger.error(
                 "Optional argument to specify buffer replacement should be 1 or 0. Provided argument list: ", args)
         else:
-            with Database() as db:
-                self.replace_with_file(self.node_id_filepath(node_id, not currently_transposed, db), replace_buffer)
+            cur_view = self.line_node_view(self.current_line_number())
+            cur_view.transposed = not cur_view.transposed
+            self.replace_with_file(self.node_id_filepath(cur_view, db), replace_buffer)
 
     @command(PluginUtils.fzf_sink_command, nargs=1, sync=True)
     def fzf_sink(self, selections: list[str]):
         for i, selected in enumerate(selections):
             # FZF adds confusing backslash before delimiter, so idx - 1
             # `017b99da-b1b5-19e9-e98d-8584cf46cfcf\^Ilaskdjf`
-            node_id = NodeId(selected[:selected.index(_FZF_LINE_DELIMITER) - 1])
-            with Database() as db:
-                node_filepath = self.node_id_filepath(node_id, False, db)
+            first_delimiter_idx = selected.index(_FZF_LINE_DELIMITER) - 1
+            node_id = NodeId(selected[:first_delimiter_idx])
+            source_id = SourceId(selected[first_delimiter_idx:selected.index(_FZF_LINE_DELIMITER,first_delimiter_idx) - 1])
+            node_filepath = self.node_id_filepath(View(node_id, source_id, None, False), db)
             if i == 0:
                 self.replace_with_file(node_filepath, False)
             else:
@@ -148,8 +166,8 @@ class Qualia(PluginDriver):
     @command("ListOrphans", sync=True)
     def list_orphans(self) -> None:
         with Database() as db:
-            orphan_fzf_lines = [fzf_input_line(node_id, db.get_node_content_lines(node_id),
-                                               False if db.get_node_descendants(node_id, True, True) else True)
+            orphan_fzf_lines = [fzf_input_line(node_id, db.get_node_content_lines(node_id, temporary),
+                                               False if db.get_node_descendants(node_id, True, True, temporary) else True)
                                 for node_id in get_orphan_node_ids(db)]
         self.fzf_run(orphan_fzf_lines, '', True)
 
@@ -159,8 +177,11 @@ class Qualia(PluginDriver):
         if not (skip_confirm or self.nvim.funcs.confirm("Remove orphans?", "&Be kind\n&Yes", 1) == 2):
             return
         with Database() as db:
-            for orphan_node_id in get_orphan_node_ids(db):
-                db.delete_node(orphan_node_id)
+            if isinstance(db, MutableDb):
+                for orphan_node_id in get_orphan_node_ids(db):
+                    db.delete_node(orphan_node_id)
+            else:
+                PluginUtils.unmutable_db_error()
 
     @function("CurrentNodeId", sync=True)
     def current_node_id(self, line_num: list[int]) -> NodeId:
